Move primary input encoding to zebra_consensus
