+++ zebra-chain/src/sapling/output.rs
@@ -105,28 +105,6 @@ impl Output {
 
         (prefix, self.zkproof)
     }
-
-    /// Encodes the primary inputs for the proof statement as 5 Bls12_381 base
-    /// field elements, to match bellman::groth16::verify_proof.
-    ///
-    /// NB: jubjub::Fq is a type alias for bls12_381::Scalar.
-    ///
-    /// https://zips.z.cash/protocol/protocol.pdf#cctsaplingoutput
-    pub fn primary_inputs(&self) -> Vec<jubjub::Fq> {
-        let mut inputs = vec![];
-
-        let cv_affine = jubjub::AffinePoint::from_bytes(self.cv.into()).unwrap();
-        inputs.push(cv_affine.get_u());
-        inputs.push(cv_affine.get_v());
-
-        let epk_affine = jubjub::AffinePoint::from_bytes(self.ephemeral_key.into()).unwrap();
-        inputs.push(epk_affine.get_u());
-        inputs.push(epk_affine.get_v());
-
-        inputs.push(self.cm_u);
-
-        inputs
-    }
 }
 
 impl OutputInTransactionV4 {
