+++ zebra-chain/src/sprout/tests.rs
@@ -3,4 +3,3 @@
 mod preallocate;
 mod test_vectors;
 mod tree;
-mod vectors;
